backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sse-starlette
openai
langchain
langchain-community
langchain-openai
langgraph
mcp
networkx
faiss-cpu
pandas
matplotlib
pypdf2
pdf2image
tavily-python
python-dotenv
pillow
sqlalchemy
langchain-tavily